# blog_system_frontend/doc/API-v2.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "包含用户注册登录、博客发布与管理、分类标签、评论及点赞功能。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/blogs": {
      "get": {
        "summary": "获取博客列表（公开）",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "可按标签名过滤，可重复添加",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "标题或内容关键词搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序字段，前缀 '-' 表示降序",
            "required": false,
            "schema": {
              "type": "string",
              "example": "-created_at"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_Blog"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "发布新博客",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogCreateRequest"
              },
              "example": {
                "title": "成功",
                "content": "我们是冠军！",
                "categoryId": 1,
                "tagIds": [
                  2
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/blogs/{blogId}": {
      "get": {
        "summary": "获取博客详情（浏览量自动 +1）",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "编辑博客（作者或管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogUpdateRequest"
              },
              "example": {
                "tagIds": [
                  1
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "删除博客（作者或管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/blogs/{blogId}/comments": {
      "get": {
        "summary": "获取某博客的评论列表（含一级回复）",
        "deprecated": false,
        "description": "",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_Comment"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "发表顶级评论",
        "deprecated": false,
        "description": "",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              },
              "example": {
                "content": "5个总冠军"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "发表成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/comments/{commentId}": {
      "delete": {
        "summary": "删除评论（评论者、博客作者或管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "7",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/blogs": {
      "get": {
        "summary": "管理员获取全站博客列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "可按标签名过滤，可重复添加",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "标题或内容关键词搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序字段，前缀 '-' 表示降序",
            "required": false,
            "schema": {
              "type": "string",
              "example": "-created_at"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_Blog"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/blogs/{blog_id}": {
      "put": {
        "summary": "管理员编辑任意博客",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 4,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogUpdateRequest"
              },
              "example": {
                "title": "成功在哪？"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "管理员删除任意博客",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/comments": {
      "get": {
        "summary": "管理员获取全站评论列表",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/comments/{comment_id}": {
      "delete": {
        "summary": "管理员删除任意评论",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "23",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri"
          },
          "bio": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Blog": {
        "type": "object",
        "properties": {
          "blog_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "view_count": {
            "type": "integer"
          },
          "like_count": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BlogCreateRequest": {
        "type": "object",
        "required": [
          "title",
          "content",
          "category_id"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category_id": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签名称列表"
          }
        }
      },
      "BlogUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category_id": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "tag_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "comment_id": {
            "type": "integer",
            "format": "int64"
          },
          "blog_id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "parent_comment_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "like_count": {
            "type": "integer"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CommentRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "PageResult_Blog": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Blog"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "PageResult_Comment": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未认证或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}